Re-organize shared workflows
